requirements.txt
moviepy==2.1.2
streamlit==1.45.0
edge_tts==6.1.19
fastapi==0.115.6
uvicorn==0.32.1
openai==1.56.1
faster-whisper==1.1.0
qwen-tts
flash-attn
loguru==0.7.3
google.generativeai==0.8.3
dashscope==1.20.14
g4f==0.5.2.2
azure-cognitiveservices-speech==1.41.1
redis==5.2.0
python-multipart==0.0.19
pyyaml
requests>=2.31.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
# Image similarity dependencies
transformers>=4.21.0
torch>=1.12.0
pillow>=9.0.0
flickrapi>=2.4.0
pywebview>=5.0
